refactor(studio): split the organization page into its four sections
OrganizationDetail was 430 lines, now 203. 312 of those lines were markup: four
sections of near-identical shape (a SectionCard, a button that opens a dialog, a
table) plus the dialogs.

- `organization-sections.tsx` — Members, Teams, TeamMembers, Roles. One module
  because they share that shape and props vocabulary; four files would have meant
  four identical import headers for no gain.
- `organization-primitives.tsx` — `SectionCard`, `ManagedTable`, `Column`, and
  `DialogState`: the shell and vocabulary the page owns as a whole.

The sections no longer touch the client. Each states intent —
`onRemoveMember`, `onRemoveTeamMember`, `onConfirmDeleteTeam`,
`onConfirmDeleteRole`, `onDialog` — and the page keeps the call, the busy flag,
and the error handling in one place. The conditional sections take their guard as
an early return rather than returning a JSX expression.

Suppression deleted. 34 auth tests green.

Refs #230.
